pyproject.toml
[tool.poetry]
name = "smart-sleep"
version = "0.1.0"
description = ""
authors = ["DanielNduati <dannduati2@gmail.com>"]
readme = "README.md"
packages = [{include = "smart_sleep"}]

[tool.poetry.dependencies]
python = "^3.10"
attrs = "21.4.0"
black = "22.6.0"
certifi = "2022.6.15"
charset-normalizer = "2.1.0"
click = "8.1.3"
flake8 = "4.0.1"
idna = "3.3"
iniconfig = "1.1.1"
isort = "5.10.1"
mccabe = "0.6.1"
mypy-extensions = "0.4.3"
packaging = "21.3"
pathspec = "0.9.0"
platformdirs = "2.5.2"
pluggy = "1.0.0"
py = "1.11.0"
pycodestyle = "2.8.0"
pyflakes = "2.4.0"
pyparsing = "3.0.9"
pytest = "7.1.2"
python-dotenv = "0.20.0"
requests = "2.28.1"
tomli = "2.0.1"
urllib3 = "1.26.11"
bleak = "^0.19.5"
aiohttp = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
